chore(automata-redux): prepare for testing
